backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
python-jose[cryptography]
bcrypt
python-dotenv
python-multipart
